Russian Location Added(2)
