+++ app/build.gradle
@@ -3,9 +3,8 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 21
     buildToolsVersion "21.1.2"
-
     defaultConfig {
-        applicationId "com.example.bogdan.dropboxphoto"
+        applicationId 'com.spb.kbv.dropboxphoto'
         minSdkVersion 19
         targetSdkVersion 21
         versionCode 1
@@ -17,6 +16,8 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    productFlavors {
+    }
 }
 
 dependencies {
@@ -25,6 +26,4 @@ dependencies {
     compile files('libs/dropbox-android-sdk-1.6.3.jar')
     compile 'com.squareup:otto:1.3.6'
     compile 'com.android.support:recyclerview-v7:+'
-
-
 }
