--- app/src/main/java/com/example/bogdan/dropboxphoto/services/Module.java
@@ -1,9 +1,0 @@
-package com.example.bogdan.dropboxphoto.services;
-
-import com.example.bogdan.dropboxphoto.infrastructure.DbxApplication;
-
-public class Module {
-    public static void register (DbxApplication application) {
-        new AccountService(application);
-    }
-}
