+++ app/src/main/java/com/spb/kbv/dropboxphoto/services/Module.java
@@ -1,0 +1,9 @@
+package com.spb.kbv.dropboxphoto.services;
+
+import com.spb.kbv.dropboxphoto.infrastructure.DbxApplication;
+
+public class Module {
+    public static void register (DbxApplication application) {
+        new AccountService(application);
+    }
+}
