+++ app/src/main/res/layout/activity_preview_photo.xml
@@ -13,7 +13,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_below="@id/include_toolbar">
-        <com.example.bogdan.dropboxphoto.views.TouchImageView
+        <com.spb.kbv.dropboxphoto.views.TouchImageView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:id="@+id/imageView"
